+++ app/src/main/java/com/vishaltelangre/nerdcalci/ui/settings/AboutSettingsScreen.kt
@@ -1,0 +1,144 @@
+package com.vishaltelangre.nerdcalci.ui.settings
+
+import android.content.Intent
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.AccountCircle
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.viewinterop.AndroidView
+import android.widget.ImageView
+import com.vishaltelangre.nerdcalci.R
+import com.vishaltelangre.nerdcalci.core.Constants
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun AboutSettingsScreen(
+    appVersion: String,
+    onBack: () -> Unit
+) {
+    val context = LocalContext.current
+
+    fun openUrl(url: String) {
+        val intent = Intent(Intent.ACTION_VIEW, android.net.Uri.parse(url))
+        context.startActivity(intent)
+    }
+
+    Scaffold(
+        topBar = {
+            CenterAlignedTopAppBar(
+                title = { Text("About") },
+                navigationIcon = {
+                    IconButton(onClick = onBack) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, "Back")
+                    }
+                },
+                colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
+                    containerColor = MaterialTheme.colorScheme.background
+                )
+            )
+        },
+        containerColor = MaterialTheme.colorScheme.background
+    ) { padding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(padding)
+                .verticalScroll(rememberScrollState())
+                .padding(vertical = 8.dp),
+            horizontalAlignment = Alignment.Start
+        ) {
+            // App Logo and Version (Centered)
+            Column(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.spacedBy(16.dp)
+            ) {
+                AndroidView(
+                    factory = { ctx ->
+                        ImageView(ctx).apply {
+                            setImageDrawable(ctx.packageManager.getApplicationIcon(ctx.packageName))
+                            scaleType = ImageView.ScaleType.FIT_CENTER
+                        }
+                    },
+                    modifier = Modifier
+                        .size(96.dp)
+                        .clip(MaterialTheme.shapes.medium)
+                )
+
+                val clipboardManager = androidx.compose.ui.platform.LocalClipboardManager.current
+                val context = LocalContext.current
+                Column(
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                    modifier = Modifier.clickable {
+                        clipboardManager.setText(androidx.compose.ui.text.AnnotatedString(appVersion))
+                        android.widget.Toast.makeText(context, "Version copied to clipboard", android.widget.Toast.LENGTH_SHORT).show()
+                    }
+                ) {
+                    Text(
+                        text = "NerdCalci",
+                        style = MaterialTheme.typography.headlineMedium,
+                        fontWeight = FontWeight.Bold,
+                        color = MaterialTheme.colorScheme.onSurface
+                    )
+                    Text(
+                        text = "Version $appVersion",
+                        style = MaterialTheme.typography.bodyMedium,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+            }
+
+            HorizontalDivider(
+                modifier = Modifier.padding(horizontal = 32.dp, vertical = 8.dp),
+                color = MaterialTheme.colorScheme.outlineVariant
+            )
+
+            // Developer Info
+            SettingsItem(
+                icon = Icons.Default.AccountCircle,
+                title = Constants.DEVELOPER_NAME,
+                subtitle = "Creator of NerdCalci",
+                onClick = { openUrl(Constants.DEVELOPER_TWITTER_URL) }
+            )
+
+            // Buy me a coffee badge
+            Image(
+                painter = painterResource(id = R.drawable.buy_me_coffee),
+                contentDescription = "Buy me a coffee",
+                modifier = Modifier
+                    .padding(top = 16.dp, start = 16.dp)
+                    .height(50.dp)
+                    .clickable {
+                        openUrl(Constants.BUY_ME_COFFEE_URL)
+                    }
+            )
+        }
+    }
+}
